+++ crates/miden-proving-service-client/src/lib.rs
@@ -13,12 +13,7 @@ extern crate std;
 use thiserror::Error;
 
 #[cfg(feature = "tx-prover")]
-pub mod generated;
-
-#[cfg(feature = "tx-prover")]
-mod prover;
-#[cfg(feature = "tx-prover")]
-pub use prover::RemoteTransactionProver;
+pub mod tx_prover;
 
 /// Protobuf definition for the Miden proving service
 pub const SERVICE_PROTO: &str = include_str!("../proto/api.proto");